requirements.txt
anyio==4.6.0
attrs==23.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Flask==3.1.0
fonttools==4.54.1
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.6
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.7
lxml==5.2.1
MarkupSafe==3.0.2
matplotlib==3.9.2
mysql-connector-python==9.2.0
numpy==2.1.2
outcome==1.3.0.post0
packaging==24.0
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.7
requests==2.31.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
webdriver-manager==4.0.1
Werkzeug==3.1.3
wsproto==1.2.0
